backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-multipart==0.0.9
python-dotenv==1.0.1
supabase==2.4.2
pytesseract==0.3.10
Pillow==10.3.0
pdf2image==1.17.0
requests==2.32.3
pydantic==2.7.1
